+++ Assets/Scripts/Note_Scripts/Note_SOs/Note_E_Block.cs
@@ -9,10 +9,10 @@ public class Note_E_Block : Note_SO
     [SerializeField] private int upgradeBlockTime;
     public override void Apply(Note_Effect_Handler noteEffectHandler, Note_SO note)
     {
-        noteEffectHandler.SetBlock((int)GetTotalStat(Game_Manager.instance.statsManager.GetStackCount(note)));
+        noteEffectHandler.SetBlock(GetTotalStat(Game_Manager.instance.statsManager.GetStackCount(note)));
     }
 
-    public override float GetTotalStat(int ownedCount)
+    public override int GetTotalStat(int ownedCount)
     {
         return blockTime + upgradeBlockTime * (ownedCount - 1);
     }
